+++ Moco/Moco/MocoUtilities.cpp
@@ -34,7 +34,6 @@
 #include <OpenSim/Simulation/Control/PrescribedController.h>
 #include <OpenSim/Simulation/Manager/Manager.h>
 #include <OpenSim/Simulation/Model/Model.h>
-#include <OpenSim/Simulation/SimbodyEngine/WeldJoint.h>
 #include <OpenSim/Simulation/StatesTrajectory.h>
 #include <OpenSim/Simulation/StatesTrajectoryReporter.h>
 
@@ -81,6 +80,10 @@ SimTK::Vector OpenSim::interpolate(const SimTK::Vector& x,
         const SimTK::Vector& y, const SimTK::Vector& newX,
         const bool ignoreNaNs) {
 
+    OPENSIM_THROW_IF(x.size() != y.size(), Exception,
+            format("Expected size of x to equal size of y, but size of x "
+                      "is %i and size of y is %i.", x.size(), y.size()));
+
     // Create vectors of non-NaN values if user set 'ignoreNaNs' argument to
     // 'true', otherwise throw an exception. If no NaN's are present in the
     // provided data vectors, the '*_no_nans' variables below will contain
@@ -96,6 +99,10 @@ SimTK::Vector OpenSim::interpolate(const SimTK::Vector& x,
             y_no_nans.push_back(y[i]);
         }
     }
+
+    OPENSIM_THROW_IF(x_no_nans.empty(), Exception,
+            "Input vectors are empty (perhaps after removing NaNs).");
+
 
     PiecewiseLinearFunction function(
             (int)x_no_nans.size(), &x_no_nans[0], &y_no_nans[0]);
@@ -610,8 +617,8 @@ MocoTrajectory OpenSim::createPeriodicTrajectory(
                 // entire name.
                 if (std::regex_match(name, regex)) {
                     matched = true;
-                    const double& oldInit = oldTraj.col(i)[0];
-                    const double& oldFinal = oldTraj.col(i)[oldN - 1];
+                    const double& oldInit = oldTraj.getElt(0, i);
+                    const double& oldFinal = oldTraj.getElt(oldN - 1, i);
                     newTraj.updBlock(oldN, i, oldN - 1, 1) =
                             oldTraj.block(1, i, oldN - 1, 1);
                     newTraj.updBlock(oldN, i, oldN - 1, 1).updCol(0) +=
@@ -624,7 +631,7 @@ MocoTrajectory OpenSim::createPeriodicTrajectory(
                 const auto regex = std::regex(pattern);
                 if (std::regex_match(name, regex)) {
                     matched = true;
-                    const double& oldFinal = oldTraj.col(i)[oldN - 1];
+                    const double& oldFinal = oldTraj.getElt(oldN - 1, i);
                     newTraj.updBlock(oldN, i, oldN - 1, 1) =
                             SimTK::Matrix(oldTraj.block(1, i, oldN - 1, 1).negate());
                     newTraj.updBlock(oldN, i, oldN - 1, 1).updCol(0) +=