+++ Moco/Tests/testConstraints.cpp
@@ -796,7 +796,8 @@ void testDoublePendulumPrescribedMotion(MocoSolution& couplerSolution,
     // state values are also set for the controls and adjuncts as dummy data.
     const auto& statesTimes = splineStateValues.getIndependentColumn();
     SimTK::Vector time((int)statesTimes.size(), statesTimes.data(), true);
-    auto mocoIterSpline = MocoTrajectory(time, splineStateValues.getColumnLabels(),
+    auto mocoIterSpline = MocoTrajectory(time,
+            splineStateValues.getColumnLabels(),
             splineStateValues.getColumnLabels(),
             splineStateValues.getColumnLabels(), {},
             splineStateValues.getMatrix(), splineStateValues.getMatrix(),
@@ -1047,16 +1048,11 @@ public:
 };
 
 template <typename TestType>
-void testDoublePendulumPointOnLineJointReaction(
-        bool enforce_constraint_derivatives, std::string dynamics_mode) {
+void testDoublePendulumJointReactionGoal(std::string dynamics_mode) {
     MocoStudy study;
-    study.setName("double_pendulum_point_on_line");
+    study.setName("double_pendulum");
     MocoProblem& mp = study.updProblem();
-    // Create double pendulum model and add the point-on-line constraint. The
-    // constraint consists of a vertical line in the y-direction (defined in
-    // ground) and the model end-effector point (the origin of body "b1").
-    // TODO: Choose a function with acceleration to ensure that accelerations
-    // are propagated successfully.
+    // Create double pendulum model and add prescribed motion constraints.
     const auto pi = SimTK::Pi;
     auto model = createDoublePendulumModel();
     model->updCoordinateSet().get("q0").setPrescribedFunction(
@@ -1099,15 +1095,14 @@ void testDoublePendulumPointOnLineJointReaction(
     ms.set_optim_solver("ipopt");
     ms.set_optim_convergence_tolerance(1e-6);
     ms.set_transcription_scheme("hermite-simpson");
-    ms.set_enforce_constraint_derivatives(enforce_constraint_derivatives);
+    ms.set_enforce_constraint_derivatives(true);
     ms.set_minimize_lagrange_multipliers(true);
     ms.set_lagrange_multiplier_weight(10);
     ms.set_multibody_dynamics_mode(dynamics_mode);
     ms.setGuess("bounds");
 
     MocoSolution solution = study.solve().unseal();
-    solution.write(
-            "testConstraints_testDoublePendulumPointOnLineJointReaction.sto");
+    solution.write("testConstraints_testDoublePendulumJointReactionGoal.sto");
 
     // Check that the actuator "push" is hitting its upper bound.
     CHECK(solution.getControl("/push")[0] == Approx(20).epsilon(1e-4));
@@ -1117,16 +1112,212 @@ void testDoublePendulumPointOnLineJointReaction(
 }
 
 TEMPLATE_TEST_CASE(
-        "DoublePendulumPointOnLineJointReaction with constraint derivatives",
+        "DoublePendulumPointJointReactionGoal with constraint derivatives",
         "[explicit]", MocoTropterSolver, MocoCasADiSolver) {
-    testDoublePendulumPointOnLineJointReaction<TestType>(true, "explicit");
+    testDoublePendulumJointReactionGoal<TestType>("explicit");
 }
 
-TEMPLATE_TEST_CASE("DoublePendulumPointOnLineJointReaction implicit with "
+TEMPLATE_TEST_CASE("DoublePendulumJointReactionGoal implicit with "
                    "constraint derivatives",
         "[implicit]", MocoCasADiSolver) {
-    testDoublePendulumPointOnLineJointReaction<TestType>(true, "implicit");
+    testDoublePendulumJointReactionGoal<TestType>("implicit");
 }
+
+struct AccelerationsAndJointReaction {
+    SimTK::Vector udot;
+    SimTK::SpatialVec reaction;
+};
+
+// TODO: Don't have direct access to multipliers. We need to know how to
+// translate multipliers into joint reactions.
+class MocoMultiplierAccelerationGoal : public MocoGoal {
+    OpenSim_DECLARE_CONCRETE_OBJECT(MocoMultiplierAccelerationGoal, MocoGoal);
+
+public:
+    MocoMultiplierAccelerationGoal(std::string jointName,
+            std::shared_ptr<AccelerationsAndJointReaction> data)
+            : m_jointName(jointName), m_data(data) {}
+    void initializeOnModelImpl(const Model&) const override {
+        setNumIntegralsAndOutputs(0, 1);
+    }
+    void calcGoalImpl(
+            const GoalInput& input, SimTK::Vector& cost) const override {
+        const auto& state = input.final_state;
+        getModel().realizeAcceleration(state);
+        m_data->udot = input.final_state.getUDot();
+        const auto& joint = getModel().getComponent<Joint>(m_jointName);
+        m_data->reaction =
+                joint.calcReactionOnChildExpressedInGround(input.final_state);
+
+        cost[0] = 0;
+    }
+
+private:
+    std::string m_jointName;
+    std::shared_ptr<AccelerationsAndJointReaction> m_data;
+};
+
+// Normally, Simbody computes its own generalized accelerations and Lagrange
+// multipliers. In Moco, when a model has kinematic constraints, we provide
+// separate multipliers. When using implicit dynamics, we provide separate
+// accelerations. This test ensures that these accelerations and multipliers are
+// those that Moco specified, not what Simbody computes.
+TEST_CASE("Goals use Moco-defined accelerations and multipliers") {
+    std::cout.rdbuf(LogManager::cout.rdbuf());
+    std::cerr.rdbuf(LogManager::cerr.rdbuf());
+    MocoStudy study;
+    study.setName("mass_welded");
+    MocoProblem& problem = study.updProblem();
+    // Rigid body welded to ground.
+    Model model;
+    const double mass = 1.0;
+    auto* body = new Body("body", mass, SimTK::Vec3(0), SimTK::Inertia(1));
+    model.addBody(body);
+
+    auto* joint = new PlanarJoint("joint", model.getGround(), *body);
+    joint->updCoordinate(PlanarJoint::Coord::RotationZ).setName("rz");
+    joint->updCoordinate(PlanarJoint::Coord::TranslationX).setName("tx");
+    joint->updCoordinate(PlanarJoint::Coord::TranslationY).setName("ty");
+    model.addJoint(joint);
+
+    auto* constr = new WeldConstraint("constraint", model.getGround(),
+            SimTK::Transform(), *body, SimTK::Transform());
+    model.addConstraint(constr);
+    model.finalizeConnections();
+    problem.setModelCopy(model);
+
+    problem.setTimeBounds(0, 1);
+
+    auto testData = std::make_shared<AccelerationsAndJointReaction>();
+    problem.addGoal<MocoMultiplierAccelerationGoal>("jointset/joint", testData);
+
+    auto& solver = study.initCasADiSolver();
+    int N = 5;
+    solver.set_num_mesh_intervals(N);
+    solver.set_minimize_lagrange_multipliers(true);
+    solver.set_multibody_dynamics_mode("implicit");
+    solver.set_optim_max_iterations(0);
+    solver.set_parallel(0);
+    auto guess = solver.createGuess("bounds");
+    int Nguess = 2;
+    guess.resampleWithNumTimes(Nguess);
+
+    const double rz = 0;
+    guess.setState(
+            "/jointset/joint/rz/value", createVectorLinspace(Nguess, rz, rz));
+    const double tx = 0;
+    guess.setState(
+            "/jointset/joint/tx/value", createVectorLinspace(Nguess, tx, tx));
+    const double ty = 0;
+    guess.setState(
+            "/jointset/joint/ty/value", createVectorLinspace(Nguess, ty, ty));
+
+    const double rz_u = 2.471;
+    guess.setState("/jointset/joint/rz/speed",
+            createVectorLinspace(Nguess, rz_u, rz_u));
+    const double tx_u = -1.61;
+    guess.setState("/jointset/joint/tx/speed",
+            createVectorLinspace(Nguess, tx_u, tx_u));
+    const double ty_u = 6.0162;
+    guess.setState("/jointset/joint/ty/speed",
+            createVectorLinspace(Nguess, ty_u, ty_u));
+
+    const double lambda0 = 0.813604;
+    guess.setMultiplier(
+            "lambda_cid3_p0", createVectorLinspace(Nguess, lambda0, lambda0));
+    const double lambda1 = 1.390461;
+    guess.setMultiplier(
+            "lambda_cid3_p1", createVectorLinspace(Nguess, lambda1, lambda1));
+    const double lambda2 = 0.614711;
+    guess.setMultiplier(
+            "lambda_cid3_p2", createVectorLinspace(Nguess, lambda2, lambda2));
+    const double lambda3 = 7.246214;
+    guess.setMultiplier(
+            "lambda_cid3_p3", createVectorLinspace(Nguess, lambda3, lambda3));
+    const double lambda4 = 4.741341;
+    guess.setMultiplier(
+            "lambda_cid3_p4", createVectorLinspace(Nguess, lambda4, lambda4));
+    const double lambda5 = 0.691361;
+    guess.setMultiplier(
+            "lambda_cid3_p5", createVectorLinspace(Nguess, lambda5, lambda5));
+
+    const double rz_a = .4024742;
+    guess.setDerivative("/jointset/joint/rz/accel",
+            createVectorLinspace(Nguess, rz_a, rz_a));
+    const double tx_a = 0.351;
+    guess.setDerivative("/jointset/joint/tx/accel",
+            createVectorLinspace(Nguess, tx_a, tx_a));
+    const double ty_a = 0.601361;
+    guess.setDerivative("/jointset/joint/ty/accel",
+            createVectorLinspace(Nguess, ty_a, ty_a));
+
+    solver.setGuess(guess);
+
+    MocoSolution solution = study.solve().unseal();
+
+    // Make sure the solution contains the multipliers and derivatives we
+    // supplied, not any computed by Simbody.
+
+    CHECK(solution.getMultiplier("lambda_cid3_p0").getElt(0, 0) ==
+            Approx(lambda0));
+    CHECK(solution.getMultiplier("lambda_cid3_p1").getElt(0, 0) ==
+            Approx(lambda1));
+    CHECK(solution.getMultiplier("lambda_cid3_p2").getElt(0, 0) ==
+            Approx(lambda2));
+    CHECK(solution.getMultiplier("lambda_cid3_p3").getElt(0, 0) ==
+            Approx(lambda3));
+    CHECK(solution.getMultiplier("lambda_cid3_p4").getElt(0, 0) ==
+            Approx(lambda4));
+    CHECK(solution.getMultiplier("lambda_cid3_p5").getElt(0, 0) ==
+            Approx(lambda5));
+
+    CHECK(solution.getDerivative("/jointset/joint/rz/accel").getElt(0, 0) ==
+            Approx(rz_a));
+    CHECK(solution.getDerivative("/jointset/joint/tx/accel").getElt(0, 0) ==
+            Approx(tx_a));
+    CHECK(solution.getDerivative("/jointset/joint/ty/accel").getElt(0, 0) ==
+            Approx(ty_a));
+
+    CHECK(testData->udot[0] == Approx(rz_a));
+    CHECK(testData->udot[1] == Approx(tx_a));
+    CHECK(testData->udot[2] == Approx(ty_a));
+
+    // The constraint Jacobian is:
+    // [1 0 0
+    //  0 0 0
+    //  0 0 0
+    //  0 1 0
+    //  0 0 1
+    //  0 0 0].
+    // That is, lambda0 exerts a moment in direction rz, lambda3 exerts
+    // a force in direction tx, and lambda4 exerts a force in direction ty.
+    auto state = model.initSystem();
+    model.setStateVariableValue(state, "/jointset/joint/rz/value", rz);
+    model.setStateVariableValue(state, "/jointset/joint/tx/value", tx);
+    model.setStateVariableValue(state, "/jointset/joint/ty/value", ty);
+    model.realizePosition(state);
+    SimTK::Matrix G;
+    model.getMatterSubsystem().calcG(state, G);
+    CAPTURE(G);
+
+
+    double Mx = lambda1;
+    double My = lambda2;
+    // Mz - lambda0 = Izz * alpha_z
+    double Mz = lambda0 + 1.0 * rz_a;
+    double Rx = mass * tx_a + lambda3;
+    double g = model.get_gravity().norm();
+    double Ry = mass * ty_a + mass * g + lambda4;
+    double Rz = lambda5;
+
+    CHECK(Mx == Approx(testData->reaction[0][0]));
+    CHECK(My == Approx(testData->reaction[0][1]));
+    CHECK(Mz == Approx(testData->reaction[0][2]));
+    CHECK(Rx == Approx(testData->reaction[1][0]));
+    CHECK(Ry == Approx(testData->reaction[1][1]));
+    CHECK(Rz == Approx(testData->reaction[1][2]));
+}
+
 
 TEST_CASE("Multipliers are correct", "") {
     std::cout.rdbuf(LogManager::cout.rdbuf());
@@ -1293,7 +1484,7 @@ TEMPLATE_TEST_CASE(
         constr->addControlPath("/tau0");
         constr->setLowerBound(Constant(lowerBound));
 
-        auto& solver = study.initSolver<TestType>();
+        study.initSolver<TestType>();
         MocoSolution solution = study.solve();
         SimTK::Vector expected(solution.getNumTimes());
         expected = lowerBound;
@@ -1315,7 +1506,7 @@ TEMPLATE_TEST_CASE(
         const double upperBound = 11.236;
         constr->setUpperBound(Constant(upperBound));
 
-        auto& solver = study.initSolver<TestType>();
+        study.initSolver<TestType>();
         MocoSolution solution = study.solve();
         SimTK::Vector expected(solution.getNumTimes());
         expected = upperBound;
@@ -1343,7 +1534,7 @@ TEMPLATE_TEST_CASE(
         constr->addControlPath("/tau0");
         constr->setLowerBound(violateLower);
         constr->setEqualityWithLower(true);
-        auto& solver = study.initSolver<TestType>();
+        study.initSolver<TestType>();
         MocoSolution solution = study.solve();
         SimTK::Vector expectedV(solution.getNumTimes());
         for (int itime = 0; itime < expectedV.size(); ++itime) {